requirements.txt
accelerate==1.3.0
anaconda-anon-usage==0.5.0
annotated-types==0.6.0
archspec==0.2.3
asttokens==3.0.0
boltons==24.1.0
Brotli==1.0.9
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
conda==25.1.1
conda-anaconda-telemetry==0.1.2
conda-anaconda-tos==0.1.2
conda-content-trust==0.2.0
conda-libmamba-solver==25.1.1
conda-package-handling==2.4.0
conda_package_streaming==0.11.0
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
decorator==5.1.1
distro==1.9.0
einops==0.8.1
ema-pytorch==0.7.7
executing==2.2.0
filelock==3.13.1
fonttools==4.56.0
frozendict==2.4.2
fsspec==2024.6.1
h5py==3.12.1
huggingface-hub==0.28.1
idna==3.7
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==2.1
kiwisolver==1.4.8
libmambapy==2.0.5
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.0
menuinst==2.2.0
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pip==25.0
platformdirs==3.10.0
pluggy==1.5.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycosat==0.6.6
pycparser==2.21
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.15.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.5.2
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
torch==2.4.1+cu121
torchaudio==2.4.1+cu121
torchvision==0.19.1+cu121
tqdm==4.67.1
traitlets==5.14.3
triton==3.0.0
truststore==0.10.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
zstandard==0.23.0
